sagemaker_on_schedule/requirements.txt
aws-psycopg2==1.3.8
boto3==1.27.1
botocore==1.30.1
jmespath==1.0.1
joblib==1.3.1
numpy==1.24.4
pandas==2.0.3
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sklearn==0.0.post5
tzdata==2023.3
urllib3==1.26.16
